weather-scheduler service split to maintain correct layer dependencies
