+++ packages/api/src/application/weather-scheduler/interfaces/weather-update.abstract.ts
@@ -1,0 +1,5 @@
+import { WeatherUpdateOptions } from '../types/weather-update.options';
+
+export abstract class WeatherUpdate {
+  abstract sendUpdates(data: WeatherUpdateOptions): Promise<void>;
+}
