+++ packages/api/src/application/weather-scheduler/weather-scheduler.module.ts
@@ -1,13 +1,21 @@
 import { Module } from '@nestjs/common';
+import { ScheduleModule } from '@nestjs/schedule';
 import { NotificationsModule } from 'src/application/notifications/notifications.module';
 import { SubscriptionModule } from 'src/application/subscriptions/subscription.module';
 import { WeatherModule } from 'src/application/weather/weather.module';
 import { CacheModule } from 'src/infrastructure/cache/cache.module';
+import { WeatherUpdateService } from 'src/infrastructure/weather/weather-update.service';
+import { WeatherUpdate } from './interfaces/weather-update.abstract';
 import { WeatherSchedulerService } from './weather-scheduler.service';
-import { ScheduleModule } from '@nestjs/schedule';
 
 @Module({
   imports: [ScheduleModule.forRoot(), NotificationsModule, WeatherModule, SubscriptionModule, CacheModule],
-  providers: [WeatherSchedulerService],
+  providers: [
+    WeatherSchedulerService,
+    {
+      provide: WeatherUpdate,
+      useClass: WeatherUpdateService,
+    },
+  ],
 })
 export class WeatherSchedulerModule {}
