+++ packages/api/src/application/weather-scheduler/weather-scheduler.service.ts
@@ -1,70 +1,29 @@
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { NotificationSubjects } from 'src/application/notifications/constants/notification-subjects.enum';
-import { NotificationType } from 'src/application/notifications/constants/notification-type.enum';
-import { NotificationsServiceInterface } from 'src/application/notifications/interfaces/notifications-service.abstract';
-import { GroupSubscriptionRepository } from 'src/application/subscriptions/interfaces/subscription-repository.abstract';
-import { WeatherUpdateOptions } from 'src/application/weather-scheduler/types/weather-update.options';
 import { Frequency } from 'src/domain/subscription/frequency.vo';
-import { CacheInvalidator } from 'src/infrastructure/cache/interfaces/cache-service.interface';
-import { transformKey } from 'src/infrastructure/cache/utils/key-transformation';
-import { CachePrefixes } from 'src/infrastructure/shared/constants/cache-prefixes.enum';
-import { WeatherProvider } from '../weather/interfaces/weather-provider.abstract';
+import { WeatherUpdate } from './interfaces/weather-update.abstract';
 
 @Injectable()
 export class WeatherSchedulerService {
-  constructor(
-    private readonly notificationsService: NotificationsServiceInterface,
-    private readonly weatherService: WeatherProvider,
-    private readonly subscriptionRepository: GroupSubscriptionRepository,
-    private readonly cacheService: CacheInvalidator,
-  ) {}
+  constructor(private readonly weatherUpdateService: WeatherUpdate) {}
 
   @Cron(CronExpression.EVERY_HOUR)
   private async sendHourlyUpdates() {
     console.log(NotificationSubjects.WEATHER_UPDATES_HOURLY);
-    await this.sendUpdates({ frequency: Frequency.HOURLY, subject: NotificationSubjects.WEATHER_UPDATES_HOURLY, invalidateCache: true });
+    await this.weatherUpdateService.sendUpdates({
+      frequency: Frequency.HOURLY,
+      subject: NotificationSubjects.WEATHER_UPDATES_HOURLY,
+      invalidateCache: true,
+    });
   }
 
   @Cron(CronExpression.EVERY_DAY_AT_8AM)
   private async sendDailyUpdates() {
     console.log(NotificationSubjects.WEATHER_UPDATES_DAILY);
-    await this.sendUpdates({ frequency: Frequency.DAILY, subject: NotificationSubjects.WEATHER_UPDATES_DAILY, invalidateCache: false });
-  }
-
-  private async sendUpdates({ frequency, subject, invalidateCache }: WeatherUpdateOptions) {
-    const grouped = await this.subscriptionRepository.findGroupedByCities(frequency);
-
-    if (invalidateCache) {
-      const cities = grouped.map((group) => group.city);
-      this.invalidateCachedWeather(cities);
-    }
-
-    for (const group of grouped) {
-      const city = group.city;
-      const weather = await this.weatherService.getCurrentWeather(city);
-
-      await Promise.all(
-        group.subscriptions.map((subscription) => {
-          this.notificationsService.sendWeatherUpdateNotification(
-            {
-              to: subscription.email,
-              subject,
-              data: {
-                city,
-                ...weather,
-              },
-            },
-            NotificationType.EMAIL,
-          );
-        }),
-      );
-    }
-  }
-
-  // invalidates weather cache
-  private async invalidateCachedWeather(cities: string[]): Promise<void> {
-    const keys = cities.map((city) => transformKey(CachePrefixes.CURRENT_WEATHER, city));
-    await this.cacheService.mdel(keys);
+    await this.weatherUpdateService.sendUpdates({
+      frequency: Frequency.DAILY,
+      subject: NotificationSubjects.WEATHER_UPDATES_DAILY,
+    });
   }
 }
